backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the Hyperclast app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg==0.30.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "channels==4.0.0",
    "channels-redis==4.1.0",
    "charset-normalizer==3.4.2",
    "cryptography==45.0.4",
    "django==5.2.3",
    "django-allauth==65.9.0",
    "django-anymail==13.0",
    "django-cors-headers>=4.9.0",
    "django-extensions==4.1",
    "django-hijack==3.7.1",
    "django-ninja==1.4.3",
    "django-rq==3.0.1",
    "gunicorn>=23.0.0",
    "html2text==2025.4.15",
    "idna==3.10",
    "litellm>=1.80.7",
    "markdown2>=2.5.4",
    "oauthlib==3.2.2",
    "pgvector>=0.4.1",
    "premailer==3.10.0",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "pycparser==2.22",
    "pycrdt-websocket[django]>=0.16.0",
    "pydantic[email]>=2.12.3",
    "pyjwt==2.10.1",
    "python-decouple==3.8",
    "redis==6.2.0",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rq==2.4.0",
    "sentry-sdk==2.30.0",
    "stripe==12.2.0",
    "urllib3==2.4.0",
    "uvicorn[standard]==0.32.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "daphne==4.1.0",
    "django-debug-toolbar==5.2.0",
    "factory-boy==3.3.3",
    "faker==37.4.0",
    "ipython==9.3.0",
    "pre-commit==4.2.0",
    "tblib==3.1.0",
    "werkzeug==3.1.3",
]
